Allow legacy HHVM to fail in Travis CI config
HHVM decided to drop support for PHP, so it's questionable if supporting
it in the future is worth the effort.
https://hhvm.com/blog/2019/02/11/hhvm-4.0.0.html

This PR does not aim to discuss how useful HHVM is and whether or not we
want to continue supporting it. It only aims to make the test output
consistent by ignoring legacy HHVM fails in the build matrix to not mark
the complete test suite as failed.

Refs https://github.com/clue/php-socket-raw/pull/42 and others.